webex-api-session/pictomaker-lite/requirements.txt
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cymem==2.0.6
en-core-web-sm==3.2.0
es-core-news-sm==3.2.0
Flask==2.0.3
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.0
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.6
nltk==3.7
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pathy==0.6.1
Pillow==9.0.1
pl-core-news-sm==3.2.0
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
regex==2022.3.15
requests==2.27.1
six==1.16.0
smart-open==5.2.1
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
SQLAlchemy==1.4.32
srsly==2.4.2
thinc==8.0.15
tqdm==4.63.1
typer==0.4.0
typing-extensions==4.1.1
urllib3==1.26.9
wasabi==0.9.0
Werkzeug==2.0.3
